Adding tests
Adding tests for the new flag. We re-run all the other tests with the
new flag (the imports at the top of the new test file). I had to modify
the function g() in test_matcher to match the name r() from the other files.
The new test (test_no_multiply_numbers) currently doesn't work though.